requirements.txt
asgiref==3.5.1
black==22.3.0
click==8.1.3
Django==3.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.3
Markdown==3.3.7
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
zipp==3.8.0
